+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 - requires Terraform 0.12.x (and AWS provider 2.x)
 - vpc.* outputs renamed to vpc_*
+- use for_each instead of count
 - rdns-forwarder: use ec2_metadata_facts instead of deprecated alias ec2_facts